% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_h5.R
\name{BPCE_listToS4}
\alias{BPCE_listToS4}
\title{Transform a list to an S4 object}
\usage{
BPCE_listToS4(robj, S4Class = NULL)
}
\arguments{
\item{robj}{A list}

\item{S4Class}{A character specifying the target S4 class.}
}
\value{
If \code{S4Class} is \code{NULL} or an unsupported class, the
original \code{robj} will return. Otherwise, a target S4 object will return.
}
\description{
Helper function to transform a list to an object with given S4 class.
}
\details{
Currently supported S4 classes:
\itemize{
\item \code{\link[SingleCellExperiment]{LinearEmbeddingMatrix}}
\item \code{\link[GenomicRanges]{GRanges}}
\item \code{\link[S4Vectors]{SimpleList}}
\item \code{\link[GenomeInfoDb]{Seqinfo}}
}
}
