% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{BPCellsExperiment-package}
\alias{BPCellsExperiment}
\alias{BPCellsExperiment-package}
\title{BPCellsExperiment: SingleCellExperiment for BPCells}
\description{
This package defines S4 classes for storing single-cell data based on 'BPCells'.
}
\author{
\strong{Maintainer}: Yuchen Li \email{ycli1995@outlook.com}

}
\keyword{internal}
