% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SingleCellExperiment.R
\name{dietSCE}
\alias{dietSCE}
\alias{dietSCE.SingleCellExperiment}
\alias{dietSCE.ChromExperiment}
\title{Slim down a SingleCellExperiment object}
\usage{
dietSCE(object, ...)

\method{dietSCE}{SingleCellExperiment}(
  object,
  assays = NULL,
  reducedDims = NULL,
  colPairs = NULL,
  rowPairs = NULL,
  altExps = NULL,
  ...
)

\method{dietSCE}{ChromExperiment}(
  object,
  assays = NULL,
  reducedDims = NULL,
  colPairs = NULL,
  rowPairs = NULL,
  altExps = NULL,
  fragments = NULL,
  annotations = NULL,
  seqinfo = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
object}

\item{...}{Arguments passed to other metheds.}

\item{assays}{Keep a subset of \code{\link[SummarizedExperiment]{assays}}
specified here. If is \code{NA}, remove all assays.}

\item{reducedDims}{Keep a subset of dimension reductions specified here. If
is \code{NA}, remove all \code{\link[SingleCellExperiment]{reducedDims}}.}

\item{colPairs}{Keep a subset of \code{\link[SingleCellExperiment]{colPairs}}
specified here. If is \code{NA}, remove all \code{colPairs}.}

\item{rowPairs}{Similar as \code{colPairs}.}

\item{altExps}{Keep a subset of \code{\link[SingleCellExperiment]{altExps}}
specified here. If is \code{NA}, remove all \code{altExps}.}

\item{fragments}{Logical scalar specifying whether to keep
\code{\link{fragments}}.}

\item{annotations}{Logical scalar specifying whether to keep
\code{\link{annotations}}.}

\item{seqinfo}{Logical scalar specifying whether to keep
\code{\link{seqinfo}}.}
}
\description{
Keep only certain aspects of the \code{SingleCellExperiment} object.
}
