% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export_data.R
\name{exportH5AD}
\alias{exportH5AD}
\alias{exportH5AD.SingleCellExperiment}
\title{Write a SingleCellExperiment to H5AD file}
\usage{
exportH5AD(object, file, ...)

\method{exportH5AD}{SingleCellExperiment}(
  object,
  file,
  name = "/",
  X = NULL,
  layers = NULL,
  obsm = NULL,
  obsp = NULL,
  varp = NULL,
  uns = NULL,
  raw = NULL,
  overwrite = FALSE,
  gzip_level = 0L,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
object}

\item{file}{Path of the H5AD file}

\item{...}{Arguments passed to other metheds.}

\item{name}{Name of the HDF5 group where to save \code{object}.}

\item{X}{Which \code{\link[SummarizedExperiment]{assay}} to be used as
`adata.X`. Default is the first one.}

\item{layers}{Which \code{\link[SummarizedExperiment]{assays}} to be used in
`adata.layers`. Default will be all assays except the `X` one.}

\item{obsm}{Which \code{\link[SingleCellExperiment]{reducedDim}} to be used
in `adata.obsm`. Default will be all reduction data.}

\item{obsp}{Which \code{\link[SingleCellExperiment]{colPair}} to be used in
`adata.obsp`.}

\item{varp}{Which \code{\link[SingleCellExperiment]{rowPair}} to be used in
`adata.varp`.}

\item{raw}{Which \code{\link[SingleCellExperiment]{altExp}} to be used in
`adata.raw`. The `nrows()` of `raw` must be larger than the `mainExp`.}

\item{overwrite}{Whether or not to overwrite the existing HDF5 link.}

\item{gzip_level}{Enable zipping at the level given here.}

\item{verbose}{Display progress.}
}
\description{
Write a SingleCellExperiment to H5AD file
}
