% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/backend.R
\name{getPath}
\alias{getPath}
\alias{getPath.IterableFragments}
\alias{getPath.IterableMatrix}
\title{Fetch the file path information for an on-disk object}
\usage{
getPath(object, ...)

\method{getPath}{IterableFragments}(object, ...)

\method{getPath}{IterableMatrix}(object, ...)
}
\arguments{
\item{object}{A file-backed object.}

\item{...}{Arguments passed to other metheds.}
}
\value{
A \code{\link{data.frame}} with three columns: path, type and group.
}
\description{
Get the file path and the link (for HDF5-backed) name for an on-disk object.
}
\details{
This function can be useful to check the file path of an object to avoid IO
conflicts.
}
\examples{
### For merged BPCells matrix
bpce <- load_example_sce()
getPath(counts(bpce))

### For merged BPCells fragments
cbpce <- load_example_csce()
getPath(counts(cbpce))
getPath(fragments(cbpce))

### For merged BPCells object
cbpce2 <- load_example_csce("unsorted")
cbpce <- merge(cbpce, cbpce2)
getPath(counts(cbpce))
getPath(fragments(cbpce))

}
