% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load-examples}
\alias{load-examples}
\alias{load_example_sce}
\alias{load_example_csce}
\alias{load_example_scme}
\title{Load the example datasets for BPCellsExperiment}
\usage{
load_example_sce(dataset = c("sorted", "unsorted"))

load_example_csce(dataset = c("sorted", "unsorted"))

load_example_scme(dataset = c("sorted", "unsorted"))
}
\arguments{
\item{dataset}{'sorted' or 'unsorted'}
}
\value{
\itemize{
\item \code{load_example_sce()} returns a \code{\link{SingleCellExperiment}}.
\item \code{load_example_csce()} returns a \code{\link{ChromExperiment}}.
\item \code{load_example_scme()} combines the two of above and returns a
\code{\link{SingleCellMultiExperiment}}
}
}
\description{
Two subsetted versions of 10x Genomics' 3k PBMC multi-omics datasets
}
\examples{
sce <- load_example_sce()
sce

csce <- load_example_csce()
csce

scme <- load_example_scme()
scme

}
