% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/backend.R
\name{openFragments}
\alias{openFragments}
\alias{openFragments.character}
\alias{openFragments.Fragment}
\alias{openFragments.IterableFragments}
\alias{showFragmentsFileMap}
\alias{clearFragmentsFileMap}
\title{Open fragments file}
\usage{
openFragments(x, group = NULL, ...)

\method{openFragments}{character}(x, group = NULL, ...)

\method{openFragments}{Fragment}(x, group = NULL, ...)

\method{openFragments}{IterableFragments}(x, group = NULL, ...)

showFragmentsFileMap()

clearFragmentsFileMap()
}
\arguments{
\item{x}{Can be one of the followings:
\itemize{
\item A \code{.tsv.gz} fragments file (must has \code{.tbi} index)
\item An HDF5 file created by \code{\link[BPCells]{write_fragments_hdf5}}.
\item A directory created by \code{\link[BPCells]{write_fragments_dir}}.
\item A \code{\link[Signac]{Fragment}} object from \pkg{Signac}.
}}

\item{group}{Control the link name of HDF5 group, where the fragments data
are stored. Default will be 'fragments'.}

\item{...}{Arguments passed to other metheds.}
}
\value{
An \code{IterableFragments} object
}
\description{
Open fragments file
}
\section{Check H5 files for opened fragments}{

When \code{openFragments} meets a \code{Fragments} object or a \code{.tsv.gz}
file, it will first check whether there is already an h5 file, which was
created by \code{\link[BPCells]{write_fragments_hdf5}} and cached the
fragments data to open. If so, it will directly open the h5 file. Otherwise,
it will calculate the \code{\link[tools]{md5sum}} of the \code{.tsv.gz}, then
cache it into a new h5 file. Use \code{showFragmentsFileMap()} to get the
file mapping table created by current R session.
}

\section{Clear temporary fragments files}{

When all needed fragments data in current R session have been stored in
proper destination paths, \code{clearFragmentsFileMap()} can be called to
delete all the temporary h5 files shown by \code{showFragmentsFileMap()}, so
that the disk storage can be saved.
}

\examples{
frags_file <- system.file(
  "extdata", "pbmc_sorted_fragments.tsv.gz",
  package = "BPCellsExperiment"
)
tmp.file <- tempfile("pbmc_sorted_fragments", fileext = ".tsv.gz")
file.copy(frags_file, tmp.file)

## Open using file name
frags <- openFragments(tmp.file)

file.copy(paste0(frags_file, ".tbi"), paste0(tmp.file, ".tbi"))

## Open using Signac::Fragment
frags <- Signac::CreateFragmentObject(tmp.file, cells = cellNames(frags))
frags <- openFragments(frags)

## Open an already existing Fragment
frags <- openFragments(frags)

}
\seealso{
\code{\link[BPCells]{open_fragments_10x}} and
\code{\link[BPCells]{open_fragments_hdf5}}
}
