% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/write_h5.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%||\%}
\alias{!!}
\alias{!!!}
\alias{:=}
\alias{\%>\%}
\alias{\%iff\%}
\alias{h5Prep}
\alias{h5Write}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{hdf5r.Extra}{\code{\link[hdf5r.Extra]{h5Prep}}, \code{\link[hdf5r.Extra]{h5Write}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{rlang}{\code{\link[rlang:dyn-dots]{:=}}, \code{\link[rlang:injection-operator]{!!}}, \code{\link[rlang:splice-operator]{!!!}}, \code{\link[rlang:op-null-default]{\%||\%}}}

  \item{SeuratObject}{\code{\link[SeuratObject:set-if-null]{\%iff\%}}}
}}

