% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMultiExperiment.R
\name{SingleCellMultiExperiment-methods}
\alias{SingleCellMultiExperiment-methods}
\alias{reducedDims,SingleCellMultiExperiment-method}
\alias{reducedDims<-,SingleCellMultiExperiment-method}
\alias{reducedDim,SingleCellMultiExperiment,Character_OR_Numeric-method}
\alias{reducedDim<-,SingleCellMultiExperiment,Character_OR_Numeric-method}
\alias{reducedDim,SingleCellMultiExperiment,missing-method}
\alias{reducedDim<-,SingleCellMultiExperiment,missing-method}
\alias{reducedDimNames,SingleCellMultiExperiment-method}
\alias{reducedDimNames<-,SingleCellMultiExperiment,character-method}
\alias{colPairs,SingleCellMultiExperiment-method}
\alias{colPairs<-,SingleCellMultiExperiment-method}
\alias{colPair,SingleCellMultiExperiment,Character_OR_Numeric-method}
\alias{colPair<-,SingleCellMultiExperiment,Character_OR_Numeric-method}
\alias{colPair,SingleCellMultiExperiment,missing-method}
\alias{colPairNames,SingleCellMultiExperiment-method}
\alias{colPairNames<-,SingleCellMultiExperiment,character-method}
\alias{path,SingleCellMultiExperiment-method}
\title{Accessing and modifying information in SingleCellMultiExperiment}
\usage{
\S4method{reducedDims}{SingleCellMultiExperiment}(x, withDimnames = TRUE, e = NULL, ...)

\S4method{reducedDims}{SingleCellMultiExperiment}(x, withDimnames = TRUE, e = NULL, ...) <- value

\S4method{reducedDim}{SingleCellMultiExperiment,Character_OR_Numeric}(x, type, withDimnames = TRUE, e = NULL, ...)

\S4method{reducedDim}{SingleCellMultiExperiment,Character_OR_Numeric}(x, type, withDimnames = TRUE, e = NULL, ...) <- value

\S4method{reducedDim}{SingleCellMultiExperiment,missing}(x, type, withDimnames = TRUE, e = NULL, ...)

\S4method{reducedDim}{SingleCellMultiExperiment,missing}(x, type, withDimnames = TRUE, e = NULL, ...) <- value

\S4method{reducedDimNames}{SingleCellMultiExperiment}(x)

\S4method{reducedDimNames}{SingleCellMultiExperiment,character}(x) <- value

\S4method{colPairs}{SingleCellMultiExperiment}(x, asSparse = FALSE, e = NULL, ...)

\S4method{colPairs}{SingleCellMultiExperiment}(x, e = NULL, ...) <- value

\S4method{colPair}{SingleCellMultiExperiment,Character_OR_Numeric}(x, type, asSparse = FALSE, e = NULL, ...)

\S4method{colPair}{SingleCellMultiExperiment,Character_OR_Numeric}(x, type, e = NULL, ...) <- value

\S4method{colPair}{SingleCellMultiExperiment,missing}(x, type, asSparse = FALSE, e = NULL, ...)

\S4method{colPairNames}{SingleCellMultiExperiment}(x, ...)

\S4method{colPairNames}{SingleCellMultiExperiment,character}(x, ...) <- value

\S4method{path}{SingleCellMultiExperiment}(object, ...)
}
\arguments{
\item{x}{A \code{\link{SingleCellMultiExperiment}} object}

\item{e}{which \code{experiment} to get or set the data. If NA, will choose
the global fields. If NULL, will choose the \code{\link{defaultExp}}.}

\item{...}{Arguments passed to other methods}

\item{value}{An object of a class specified in the S4 method signature.}

\item{type, withDimnames, asSparse}{Arguments passed to methods for
\code{SingleCellExperiment} (\code{reducedDims}, \code{colPairs} and
\code{rowPairs})}

\item{object}{A \code{\link{SingleCellMultiExperiment}} object}
}
\description{
A set of getter and setter functions to extract the single-cell specific data
fields for a \code{\link{SingleCellMultiExperiment}} object.
}
\seealso{
\itemize{
\item \code{\link[SingleCellExperiment]{reducedDims}}
\item \code{\link[SingleCellExperiment]{colPairs}}
}
}
