% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMultiExperiment.R
\name{SingleCellMultiExperiment-subset}
\alias{SingleCellMultiExperiment-subset}
\alias{[,SingleCellMultiExperiment,ANY,ANY,ANY-method}
\alias{[[<-,SingleCellMultiExperiment,ANY,ANY,ANY-method}
\alias{[<-,SingleCellMultiExperiment,ANY,ANY,ANY-method}
\alias{subsetByColData,SingleCellMultiExperiment,ANY-method}
\alias{subsetByColData,SingleCellMultiExperiment,character-method}
\alias{subsetByColumn,MultiAssayExperiment,ANY-method}
\alias{subsetByAssay,SingleCellMultiExperiment-method}
\title{Subsetting a SingleCellMultiExperiment object}
\usage{
\S4method{[}{SingleCellMultiExperiment,ANY,ANY,ANY}(x, i, j, k, ..., drop = FALSE)

\S4method{[[}{SingleCellMultiExperiment,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{SingleCellMultiExperiment,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{subsetByColData}{SingleCellMultiExperiment,ANY}(x, y)

\S4method{subsetByColData}{SingleCellMultiExperiment,character}(x, y)

\S4method{subsetByColumn}{MultiAssayExperiment,ANY}(x, y)

\S4method{subsetByAssay}{SingleCellMultiExperiment}(x, y)
}
\arguments{
\item{x}{A \code{\link{SingleCellMultiExperiment}} object.}

\item{i}{Either a \code{character}, \code{integer}, \code{logical} or
\code{GRanges} object for subsetting by rows}

\item{j}{Either a \code{character}, \code{logical}, or \code{numeric} vector
for subsetting by \code{colData} rows. See details for more information.}

\item{k}{Either a \code{character}, \code{logical}, or \code{numeric} vector
for subsetting by assays}

\item{...}{Additional arguments passed on to lower level functions.}

\item{drop}{logical (default FALSE) whether to drop all empty assay elements
in the \code{ExperimentList}}

\item{value}{An object of a class specified in the S4 method signature.}

\item{y}{Any argument used for subsetting, can be a \code{character},
\code{logical}, \code{integer}, \code{list} or \code{List} vector}
}
\description{
Extracting and dividing a \code{SingleCellMultiExperiment}. These methods
actually inherit from \code{\link[MultiAssayExperiment]{subsetBy}} in
\pkg{MultiAssayExperiment} package.
}
\seealso{
\code{\link[MultiAssayExperiment]{subsetBy}}
}
