% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/write_h5.R
\name{writeH5SCE}
\alias{writeH5SCE}
\alias{writeH5SCE.default}
\alias{writeH5SCE.SingleCellExperiment}
\alias{writeH5SCE.ChromExperiment}
\title{Write a SingleCellExperiment to HDF5 file}
\usage{
writeH5SCE(object, file, ...)

\method{writeH5SCE}{default}(object, file, ...)

\method{writeH5SCE}{SingleCellExperiment}(
  object,
  file,
  name = "/",
  overwrite = FALSE,
  gzip_level = 0L,
  verbose = TRUE,
  ...
)

\method{writeH5SCE}{ChromExperiment}(
  object,
  file,
  name = "/",
  overwrite = FALSE,
  gzip_level = 0L,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
object}

\item{file}{Path of the HDF5 file}

\item{...}{Arguments passed to other metheds.}

\item{name}{Name of the HDF5 group where to save \code{object}.}

\item{overwrite}{Whether or not to overwrite the existing HDF5 link.}

\item{gzip_level}{Enable zipping at the level given here.}

\item{verbose}{Display progress.}
}
\description{
Write a SingleCellExperiment to HDF5 file
}
