% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/write_h5.R
\name{writeH5SCME}
\alias{writeH5SCME}
\alias{writeH5SCME.default}
\alias{writeH5SCME.SingleCellMultiExperiment}
\title{Write a SingleCellMultiExperiment to HDF5 file}
\usage{
writeH5SCME(object, file, ...)

\method{writeH5SCME}{default}(object, file, ...)

\method{writeH5SCME}{SingleCellMultiExperiment}(
  object,
  file,
  name = "/",
  overwrite = FALSE,
  gzip_level = 0L,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{\link{SingleCellMultiExperiment}} object}

\item{file}{Path of the HDF5 file}

\item{...}{Arguments passed to other metheds.}

\item{name}{Name of the HDF5 group where to save \code{object}.}

\item{overwrite}{Whether or not to overwrite the existing HDF5 link.}

\item{gzip_level}{Enable zipping at the level given here.}

\item{verbose}{Display progress.}
}
\description{
Write a SingleCellMultiExperiment to HDF5 file
}
